Manage requirements with pip-compile
Add a new `requirements.in` file limited to the top-level requirements
the site and build script depend on, and compile that to a fully pinned
`requriements.txt`. [`pip-compile`] has a number of benefits; the main
one this change intends to incorporate is the removal of unused
dependencies from the requirements file.

I plan to follow this change with some additional requirements
automation.

[`pip-compile`]: https://github.com/jazzband/pip-tools#example-usage-for-pip-compile